Move Counts from model to Functionable service
Counts was a class with only class methods in app/models despite not
being an ActiveRecord model. Convert to a Functionable module in
app/services, fix STOREAGE_KEY typo, and remove unused redis methods.
